+++ plugins/catalog-backend-module-aws/src/service/AwsS3EntityProviderCatalogModule.ts
@@ -14,10 +14,10 @@
  * limitations under the License.
  */
 
+import { loggerToWinstonLogger } from '@backstage/backend-common';
 import {
   coreServices,
   createBackendModule,
-  loggerToWinstonLogger,
 } from '@backstage/backend-plugin-api';
 import { catalogProcessingExtensionPoint } from '@backstage/plugin-catalog-node';
 import { AwsS3EntityProvider } from '../providers';