+++ plugins/scaffolder-backend-module-rails/src/actions/fetch/rails/index.ts
@@ -54,7 +54,7 @@ export function createFetchRailsAction(options: {
   }>({
     id: 'fetch:rails',
     description:
-      'Downloads a template from the given URL into the workspace, and runs a rails new generator on it.',
+      'Downloads a template from the given `url` into the workspace, and runs a rails new generator on it.',
     schema: {
       input: {
         type: 'object',