+++ plugins/search-react/src/components/SearchResult/SearchResult.stories.tsx
@@ -218,3 +218,32 @@ export const GroupLayout = () => {
     </SearchResult>
   );
 };
+
+export const WithCustomNoResultsComponent = () => {
+  return (
+    <SearchResult noResultsComponent={<>No results were found</>}>
+      {({ results }) => (
+        <List>
+          {results.map(({ type, document }) => {
+            switch (type) {
+              case 'custom-result-item':
+                return (
+                  <CustomResultListItem
+                    key={document.location}
+                    result={document}
+                  />
+                );
+              default:
+                return (
+                  <DefaultResultListItem
+                    key={document.location}
+                    result={document}
+                  />
+                );
+            }
+          })}
+        </List>
+      )}
+    </SearchResult>
+  );
+};