+++ plugins/bazaar/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @backstage/plugin-bazaar
 
+## 0.2.3-next.1
+
+### Patch Changes
+
+- 0bf734dbac: Created soft cut on `ProjectCard` to indicate more content.
+- 41d0b0a24a: Added `isBazaarAvailable` helper to be used with the `EntitySwitch`.
+- Updated dependencies
+  - @backstage/cli@0.22.1-next.1
+  - @backstage/catalog-client@1.3.0-next.1
+  - @backstage/catalog-model@1.1.5-next.1
+  - @backstage/core-components@0.12.3-next.1
+  - @backstage/core-plugin-api@1.2.1-next.0
+  - @backstage/errors@1.1.4
+  - @backstage/theme@0.2.16
+  - @backstage/plugin-catalog@1.7.2-next.1
+  - @backstage/plugin-catalog-react@1.2.4-next.1
+
 ## 0.2.3-next.0
 
 ### Patch Changes