+++ plugins/catalog-backend-module-bitbucket-cloud/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @backstage/plugin-catalog-backend-module-bitbucket-cloud
 
+## 0.1.7-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-plugin-api@0.2.1-next.0
+  - @backstage/backend-common@0.18.0-next.0
+  - @backstage/config@1.0.6-next.0
+  - @backstage/plugin-catalog-backend@1.7.0-next.1
+  - @backstage/plugin-catalog-node@1.3.1-next.1
+  - @backstage/plugin-events-node@0.2.1-next.0
+  - @backstage/backend-tasks@0.4.1-next.0
+  - @backstage/catalog-client@1.3.0-next.1
+  - @backstage/catalog-model@1.1.5-next.1
+  - @backstage/integration@1.4.2-next.0
+  - @backstage/plugin-bitbucket-cloud-common@0.2.3-next.0
+  - @backstage/plugin-catalog-common@1.0.10-next.1
+
 ## 0.1.7-next.0
 
 ### Patch Changes