+++ plugins/catalog-backend-module-openapi/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @backstage/plugin-catalog-backend-module-openapi
 
+## 0.1.6-next.1
+
+### Patch Changes
+
+- f9abdb64fc: Updated internal usage of UrlReader interface.
+- Updated dependencies
+  - @backstage/backend-common@0.18.0-next.0
+  - @backstage/config@1.0.6-next.0
+  - @backstage/plugin-catalog-backend@1.7.0-next.1
+  - @backstage/plugin-catalog-node@1.3.1-next.1
+  - @backstage/catalog-model@1.1.5-next.1
+  - @backstage/integration@1.4.2-next.0
+  - @backstage/types@1.0.2
+
 ## 0.1.6-next.0
 
 ### Patch Changes