+++ plugins/config-schema/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @backstage/plugin-config-schema
 
+## 0.1.37-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/config@1.0.6-next.0
+  - @backstage/core-components@0.12.3-next.1
+  - @backstage/core-plugin-api@1.2.1-next.0
+  - @backstage/errors@1.1.4
+  - @backstage/theme@0.2.16
+  - @backstage/types@1.0.2
+
 ## 0.1.37-next.0
 
 ### Patch Changes