+++ plugins/example-todo-list-common/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @internal/plugin-todo-list-common
 
+## 1.0.8-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-permission-common@0.7.3-next.0
+
 ## 1.0.7
 
 ### Patch Changes