+++ plugins/lighthouse/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @backstage/plugin-lighthouse
 
+## 0.3.14-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/config@1.0.6-next.0
+  - @backstage/catalog-model@1.1.5-next.1
+  - @backstage/core-components@0.12.3-next.1
+  - @backstage/core-plugin-api@1.2.1-next.0
+  - @backstage/theme@0.2.16
+  - @backstage/plugin-catalog-react@1.2.4-next.1
+
 ## 0.3.14-next.0
 
 ### Patch Changes